refactor - reference 수정
